+++ CHANGELOG.rst
@@ -1,5 +1,9 @@
 Unreleased changes
 ------------------
+
+* ``h5extract`` now has the option ``--aashower-legacy`` which is needed
+  to account for the old number of aashower reco_stages which has now changed.
+
 Version 9
 ---------